OSCORE: Support for Observe functionality
Adds support for using Observe with OSCORE messages.
Also adds a JUnit test for testing Observe with OSCORE.
This commit also does some small fixes to the other OSCORE code and comments.


Small fix in Encryptor class when creating Request OSCORE option value


Updated copyright header in OSCoreObserveTest


Make sure to not call setDefaultCoapStackFactory multiple times


Changed asserts for comparing arrays and one value in OSCoreTest


Updated year in copyright header of OSCoreObserveTest


Added CoapNetworkRule to JUnit test OSCoreObserveTest

+++ cf-oscore/src/main/java/org/eclipse/californium/oscore/ResponseDecryptor.java
@@ -101,7 +101,11 @@ public class ResponseDecryptor extends Decryptor {
 		eOptions = OptionJuggle.merge(eOptions, uOptions);
 		response.setOptions(eOptions);
 
-		db.removeToken(token);
+		//Remove token after response is received, unless it has Observe
+		//If is has Observe it will be removed after cancellation elsewhere
+		if(response.getOptions().hasObserve() == false) {
+			db.removeToken(token);
+		}
 
 		return response;
 	}
